requirements.txt
absl-py
aiohttp
aiosignal
altair
annotated-types
anthropic
anyio
apptools
astunparse
attrs
beautifulsoup4
blinker
build
cachetools
certifi
charset-normalizer
click
compute_rhino3d
configobj
contourpy
cx_Freeze
cycler
dataclasses-json
distro
dmgbuild
docstring_parser
docutils
dotenv-python
ds-store
envisage
faiss-cpu
fastapi
filelock
fire
Flask
flatbuffers
fonttools
frozendict
frozenlist
gast
gitdb
GitPython
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-generativeai
google-pasta
googleapis-common-protos
grpcio
grpcio-status
h11
h5py
html5lib
httpcore
httplib2
httpx
id
idna
itsdangerous
jaraco.classes
jaraco.context
jaraco.functools
Jinja2
jiter
joblib
jsonpatch
jsonpointer
jsonschema
jsonschema-specifications
keras
keyring
kiwisolver
langchain
langchain-community
langchain-core
langchain-openai
langchain-text-splitters
langsmith
libclang
lxml
mac-alias
mangum
Markdown
markdown-it-py
MarkupSafe
marshmallow
matplotlib
mayavi
mdurl
mesh
ml-dtypes
more-itertools
MouseInfo
multidict
multitasking
mypy-extensions
namex
narwhals
nh3
numpy
numpy-stl
ollama
open3d-python
openai
opt_einsum
optree
orjson
outcome
packaging
pandas
pdf2image
peewee
pillow
platformdirs
proto-plus
protobuf
pyarrow
pyasn1
pyasn1_modules
PyAutoGUI
pydantic
pydantic-settings
pydantic_core
pydeck
pyface
pygame
PyGetWindow
Pygments
PyMsgBox
pyobjc
pyobjc-core
pyobjc-framework-Accessibility
pyobjc-framework-Accounts
pyobjc-framework-AddressBook
pyobjc-framework-AdServices
pyobjc-framework-AdSupport
pyobjc-framework-AppleScriptKit
pyobjc-framework-AppleScriptObjC
pyobjc-framework-ApplicationServices
pyobjc-framework-AppTrackingTransparency
pyobjc-framework-AudioVideoBridging
pyobjc-framework-AuthenticationServices
pyobjc-framework-AutomaticAssessmentConfiguration
pyobjc-framework-Automator
pyobjc-framework-AVFoundation
pyobjc-framework-AVKit
pyobjc-framework-AVRouting
pyobjc-framework-BackgroundAssets
pyobjc-framework-BrowserEngineKit
pyobjc-framework-BusinessChat
pyobjc-framework-CalendarStore
pyobjc-framework-CallKit
pyobjc-framework-CFNetwork
pyobjc-framework-Cinematic
pyobjc-framework-ClassKit
pyobjc-framework-CloudKit
pyobjc-framework-Cocoa
pyobjc-framework-Collaboration
pyobjc-framework-ColorSync
pyobjc-framework-Contacts
pyobjc-framework-ContactsUI
pyobjc-framework-CoreAudio
pyobjc-framework-CoreAudioKit
pyobjc-framework-CoreBluetooth
pyobjc-framework-CoreData
pyobjc-framework-CoreHaptics
pyobjc-framework-CoreLocation
pyobjc-framework-CoreMedia
pyobjc-framework-CoreMediaIO
pyobjc-framework-CoreMIDI
pyobjc-framework-CoreML
pyobjc-framework-CoreMotion
pyobjc-framework-CoreServices
pyobjc-framework-CoreSpotlight
pyobjc-framework-CoreText
pyobjc-framework-CoreWLAN
pyobjc-framework-CryptoTokenKit
pyobjc-framework-DataDetection
pyobjc-framework-DeviceCheck
pyobjc-framework-DictionaryServices
pyobjc-framework-DiscRecording
pyobjc-framework-DiscRecordingUI
pyobjc-framework-DiskArbitration
pyobjc-framework-DVDPlayback
pyobjc-framework-EventKit
pyobjc-framework-ExceptionHandling
pyobjc-framework-ExecutionPolicy
pyobjc-framework-ExtensionKit
pyobjc-framework-ExternalAccessory
pyobjc-framework-FileProvider
pyobjc-framework-FileProviderUI
pyobjc-framework-FinderSync
pyobjc-framework-FSEvents
pyobjc-framework-GameCenter
pyobjc-framework-GameController
pyobjc-framework-GameKit
pyobjc-framework-GameplayKit
pyobjc-framework-HealthKit
pyobjc-framework-ImageCaptureCore
pyobjc-framework-InputMethodKit
pyobjc-framework-InstallerPlugins
pyobjc-framework-InstantMessage
pyobjc-framework-Intents
pyobjc-framework-IntentsUI
pyobjc-framework-IOBluetooth
pyobjc-framework-IOBluetoothUI
pyobjc-framework-IOSurface
pyobjc-framework-iTunesLibrary
pyobjc-framework-KernelManagement
pyobjc-framework-LatentSemanticMapping
pyobjc-framework-LaunchServices
pyobjc-framework-libdispatch
pyobjc-framework-libxpc
pyobjc-framework-LinkPresentation
pyobjc-framework-LocalAuthentication
pyobjc-framework-LocalAuthenticationEmbeddedUI
pyobjc-framework-MailKit
pyobjc-framework-MapKit
pyobjc-framework-MediaAccessibility
pyobjc-framework-MediaLibrary
pyobjc-framework-MediaPlayer
pyobjc-framework-MediaToolbox
pyobjc-framework-Metal
pyobjc-framework-MetalFX
pyobjc-framework-MetalKit
pyobjc-framework-MetalPerformanceShaders
pyobjc-framework-MetalPerformanceShadersGraph
pyobjc-framework-MetricKit
pyobjc-framework-MLCompute
pyobjc-framework-ModelIO
pyobjc-framework-MultipeerConnectivity
pyobjc-framework-NaturalLanguage
pyobjc-framework-NetFS
pyobjc-framework-Network
pyobjc-framework-NetworkExtension
pyobjc-framework-NotificationCenter
pyobjc-framework-OpenDirectory
pyobjc-framework-OSAKit
pyobjc-framework-OSLog
pyobjc-framework-PassKit
pyobjc-framework-PencilKit
pyobjc-framework-PHASE
pyobjc-framework-Photos
pyobjc-framework-PhotosUI
pyobjc-framework-PreferencePanes
pyobjc-framework-PushKit
pyobjc-framework-Quartz
pyobjc-framework-QuickLookThumbnailing
pyobjc-framework-ReplayKit
pyobjc-framework-SafariServices
pyobjc-framework-SafetyKit
pyobjc-framework-SceneKit
pyobjc-framework-ScreenCaptureKit
pyobjc-framework-ScreenSaver
pyobjc-framework-ScreenTime
pyobjc-framework-ScriptingBridge
pyobjc-framework-SearchKit
pyobjc-framework-Security
pyobjc-framework-SecurityFoundation
pyobjc-framework-SecurityInterface
pyobjc-framework-SensitiveContentAnalysis
pyobjc-framework-ServiceManagement
pyobjc-framework-SharedWithYou
pyobjc-framework-SharedWithYouCore
pyobjc-framework-ShazamKit
pyobjc-framework-Social
pyobjc-framework-SoundAnalysis
pyobjc-framework-Speech
pyobjc-framework-SpriteKit
pyobjc-framework-StoreKit
pyobjc-framework-Symbols
pyobjc-framework-SyncServices
pyobjc-framework-SystemConfiguration
pyobjc-framework-SystemExtensions
pyobjc-framework-ThreadNetwork
pyobjc-framework-UniformTypeIdentifiers
pyobjc-framework-UserNotifications
pyobjc-framework-UserNotificationsUI
pyobjc-framework-VideoSubscriberAccount
pyobjc-framework-VideoToolbox
pyobjc-framework-Virtualization
pyobjc-framework-Vision
pyobjc-framework-WebKit
PyOpenGL
pyparsing
pyperclip
pyproject_hooks
PyQt6
PyQt6-Qt6
PyQt6-sip
PyRect
PyScreeze
PySocks
python-dateutil
python-dotenv
python-utils
pytweening
pytz
PyYAML
readme_renderer
referencing
regex
requests
requests-toolbelt
rfc3986
rhino3dm
rich
richxerox
rpds-py
rsa
rubicon-objc
scikit-learn
scipy
selenium
setuptools
shellingham
six
smmap
sniffio
sortedcontainers
soupsieve
SQLAlchemy
starlette
streamlit
tenacity
tensorboard
tensorboard-data-server
tensorflow
termcolor
threadpoolctl
tiktoken
toml
tornado
tqdm
traits
traitsui
trimesh
trio
trio-websocket
twine
typer
typing-inspect
typing-inspection
typing_extensions
tzdata
uritemplate
urllib3
uvicorn
vtk
webencodings
websocket-client
Werkzeug
wheel
wrapt
wsproto
yarl
yfinance
youtube-transcript-api